requirements.txt
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
Flag==0.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.1
python-i18n==0.3.9
python-telegram-bot==21.3
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
